+++ src/main/java/io/paradaux/friendlybot/listeners/ReadyListener.java
@@ -24,7 +24,7 @@
 package io.paradaux.friendlybot.listeners;
 
 import io.paradaux.friendlybot.FriendlyBot;
-import io.paradaux.friendlybot.utils.models.objects.DiscordEventListener;
+import io.paradaux.friendlybot.utils.models.types.DiscordEventListener;
 import net.dv8tion.jda.api.entities.Guild;
 import net.dv8tion.jda.api.events.ReadyEvent;
 import org.slf4j.Logger;
